delete redundant
