--- frontend/src/app/models/database-items/base-db-item.ts
@@ -1,4 +1,0 @@
-export interface BaseDbItem {
-    schema: string;
-    name: string;
-}
