+++ frontend/src/app/models/database-items/database-item-type-name.ts
@@ -1,0 +1,10 @@
+import { DatabaseItemType } from './database-item-type';
+
+export const DatabaseItemTypeName = {
+    [DatabaseItemType.Table]: 'Table',
+    [DatabaseItemType.Type]: 'Type',
+    [DatabaseItemType.Constraint]: 'Constraint',
+    [DatabaseItemType.StoredProcedure]: 'StoredProcedure',
+    [DatabaseItemType.Function]: 'Function',
+    [DatabaseItemType.View]: 'View',
+};
