+++ frontend/src/app/models/database-items/database-item-type.ts
@@ -6,12 +6,3 @@ export enum DatabaseItemType {
     Function,
     View,
 }
-
-export const DatabaseItemTypeName = {
-    [DatabaseItemType.Table]: 'Table',
-    [DatabaseItemType.Type]: 'Type',
-    [DatabaseItemType.Constraint]: 'Constraint',
-    [DatabaseItemType.StoredProcedure]: 'StoredProcedure',
-    [DatabaseItemType.Function]: 'Function',
-    [DatabaseItemType.View]: 'View',
-};
